backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websocket-client>=1.4.0
requests>=2.32.3
python-multipart>=0.0.9
pydantic>=2.0.0
google-genai>=1.0.0
